requirements.txt
python-dotenv

openai
tiktoken

gradio
